requirements.txt
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
chromadb>=0.4.0
numpy>=1.24.0
requests>=2.28.0
matplotlib>=3.6.0
seaborn>=0.12.0
scikit-learn>=1.3.0
umap-learn>=0.5.0
plotly>=5.15.0
pandas>=2.0.0
streamlit>=1.28.0
rq>=1.15.0
redis>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0